symbols: Restore single-regexp for files which have complex symbols
For instance, locale/fr/symbols.dic has both

in-word '       (?<=[^\W_])['’]
’       apostrophe droite

that can match ’, but the second rule is supposed not to match when the
first one does. We have to keep them in the same regexp for this to
happen.

So we refine 031eda9a45ad ("symbols: Split simple symbol tables into chunks
of 1000 lines") into splitting only for files which do not have complex
symbols, and thus for which we don't have that situation, notably
emojis.dic which is the one that has the size issue.